Remove extra ']' character in `clearLine` sequence
There was an extra ']' character that showed up at the beginning of each
progress bar line, and it was caused by a stray character after the CR
in the `clearLine` escape sequence.  This extra character is not present
in the equivalent `goprogressbar` function, and looks like it might have
been inserted by automatic brace-matching when the custom 'Gopass'
progress bar was created.