Implement `gh`-style versioning
Versions are based on Git tags using `git describe --tags`, which
provides better indication of dev builds and also reduces the number of
files that need to be updated for each new release.

Note that this is not yet fully supported by all of the release helper
utilities.  In particular, the `VERSION` file shouldn't be necessary.