Attempt to fix handling of `rcs init` command
If a store does not exist, then it should be initialized with the `init`
command.  The purpose of the `rcs init` command is to initialize a Git
repository in an existing "fs" store that has no revision control, so
we won't get the desired result if we fail the init, or if we just do
nothing, when the store is using the "fs" backend.  It also makes no
sense to check the backend for the context because that is initialized
from the root store only, and the actual storage could be different if
the "--store" option was used.