Move `gopass git` to the regular commands
Now that Git is the only RCS provider, `gopass git` is the only extra
command introduced by the storage backends.  We can eliminate the extra
complexity of the extension commands interface by moving it to the
regular commands array, and there's even a leftover `Git` command func
that we can reuse.