Remove overly-opinionated git config options
Some git settings should be left to the user, and the user's default
settings are usually just fine.  Config options in the local repository
have the highest precedence, so they will override any other preferences
configured by the user.  `gopass` should not be enforcing specific
settings that the user can't change, especially when an FAQ entry is
needed to explain potential problems with an enforced setting.