Remove the `env` command
An environment variable for a single secret can be set just as easily
using command substitution.  The value of setting variables for all
secrets in a directory seems a bit dubious, especially since multiple
secrets with the same name in different sub-directories will collide on
the same variable name, so a value may not come from the secret you
expect.