Remove the `convert` command
We only have one revision control provider now, so conversion is really
just changing the encryption.  This is relatively easy to implement as
a shell script, similar to the sample `passage` migration script shown
in <https://github.com/FiloSottile/passage/blob/main/README>.  The
`convert` command would probably be better implemented as a separate
utility using `gopass` as a library (similar to `gopass-hibp`).