Remove desktop notifications
As a command-line utility, most operations won't be running in the
background even if they might take some time, so a desktop notification
shouldn't be necessary.  If commands are being scripted then it should
be up to the script to determine if notifications are needed.  The only
built-in desktop notification that really made any sense was for the
delayed clearing of the clipboard, but it doesn't seem worth the trouble
just for that one case.