Move the "rcs init" subcommand to the "git" command
The "init" subcommand was the only subcommand left under the old "rcs"
command.  Since it runs a `git init` command, but with important
additional configuration, we can make it a subcommand of "gopass git"
so that it will be used instead of just forwarding the "init" command
directly to `git`.  This allows the "git" command to completely replace
the "rcs" command.