Use simple progress spinner for `sync` messages
The spinner indicates activity during the push-pull (if it is needed).
Import/export of missing keys generally adds very little to the duration
of the `sync` operation, so it doesn't really need a separate message
unless it fails.  This makes the output much less cluttered, which is
especially welcome for auto-sync.

The indication of added and removed secrets is also removed.  There is
nothing wrong with the added/removed counts themselves, but the "no
changes" message could be misleading because it ignores modifications
to existing secrets.  In most cases, the user probably doesn't care if
the contents of the store changed, but only that the local store is
up-to-date, so no indication at all seems better than a potentially
misleading one.  This also eliminates the overhead, however small, of
gathering a complete list of all secrets in the store _twice_.