Remove automatic environment checks
This doesn't actually do anything except on macOS, and the only check on
macOS is to look for an "insecure" configuration of `pinentry-mac`.  GPG
encryption is too complex for an automated process to safely hold the
hand of an inexperienced user, and an experienced GPG user should be
trusted to have already configured their GPG environment appropriately
based on the level of security that they are comfortable with.