Remove `safecontent` setting
As stated in the docs, the `safecontent` setting is not completely
reliable, and even the original author has considered removing it.

It is a standardized convention to store the password in the first line
of a secret, but other values in the secret could be just as sensitive.
A user should always act as if the entire content of the secret is
unsafe, and the `safecontent` setting could give a false sense of
security.

It's also not compatible with the command-line interface of the original
`pass` utility.