Remove `AskForGitConfigUser` function
The implementation of this function is broken, and it will **always**
return empty strings for both name and email.  Based on the Git history,
it has probably been broken for quite some time, and apparently no-one
has missed it.  It only handles GPG keys anyway.

It could be fixed by using the correct `FormatKey` template strings, but
it still has several rather annoying and non-intuitive behaviors:
- Instead of the selection list used for the recipient key, it issues a
  yes/no prompt for each identity until you accept one of them.
- Accepting an identity doesn't mean that the exact name and email will
  be used for the Git config.  Either of these values could be silently
  ignored if `DetectName` or `DetectEmail` already found a non-empty
  value in an environment variable.
- If a value _is_ needed from the selected identity, it _still_ isn't
  used automatically.  It just becomes a default value, and you still
  have to respond to another prompt to accept the default.